refactor: 대시보드 코드스타일 통일
